@@ -17,8 +17,6 @@ class String
   end
 
   def sanitize_unicode
-    self.encode("UTF-8", { :undef => :replace,
-                           :invalid => :replace,
-                           :replace => "?" }).gsub("\u0000","")
+    self.encode("UTF-8", undef: :replace, invalid: :replace, replace: "?" ).gsub("\u0000","")
   end
 end
