@@ -111,8 +111,8 @@ module Model
       else # just the light version
         {
           "id" => self.id,
-          "name" =>  URI.escape(self.name),
-          "task_name" => URI.escape(self.task_name),
+          "name" =>  URI.encode_www_form_component(self.name),
+          "task_name" => URI.encode_www_form_component(self.task_name),
           "timestamp_start" => self.timestamp_start,
           "timestamp_end" => self.timestamp_end,
           "options" => self.options,
@@ -125,8 +125,8 @@ module Model
       {
         "id" => self.id,
         "job_id" => self.job_id,
-        "name" =>  URI.escape(self.name),
-        "task_name" => URI.escape(self.task_name),
+        "name" =>  URI.encode_www_form_component(self.name),
+        "task_name" => URI.encode_www_form_component(self.task_name),
         "timestamp_start" => self.timestamp_start,
         "timestamp_end" => self.timestamp_end,
         "project" => self.project.name,
