@@ -51,7 +51,7 @@ class Domain < Intrigue::Task::BaseTask
 
       # grab whois info & all nameservers
       if soa_details
-        out = whois(lookup_name)
+        out = whois(lookup_name).first
         if out
           _set_entity_detail("whois_full_text", out["whois_full_text"])
           _set_entity_detail("contacts", out["contacts"])
