@@ -31,7 +31,7 @@ class NetBlock < Intrigue::Task::BaseTask
       _log "Skipping lookup, we already have the details"
       out = @entity.details
     else # do the lookup
-      out = whois(lookup_string) || {}
+      out = whois(lookup_string).first
       # make sure not to overwrite the name in the details
       out = out.merge({"name" => netblock_string, "_hidden_name" => netblock_string})
       # lazy but easier than setting invidually
