@@ -8,6 +8,8 @@ module Whois
 
   def whois(lookup_string)
 
+    out = []
+
     begin
       whois = ::Whois::Client.new #(:timeout => 60)
       answer = whois.lookup(lookup_string)
@@ -42,34 +44,51 @@ module Whois
     # grab the parser so we can get specific fields
     parser = answer.parser
 
-    out = {}
-    out["whois_full_text"] = "#{answer.content}".force_encoding('ISO-8859-1').sanitize_unicode
+    # store whois text so we can get to the right RIR
+    whois_text = "#{answer.content}".force_encoding('ISO-8859-1').sanitize_unicode
+    
+    ###
+    ### Note that we need to handle the potential for multiple ranges here, since a 
+    ### RIPE range can be split into many CIDRs
+    ###
 
     if lookup_string.is_ip_address?
+
       # Handle this at the regional internet registry level
-      if out["whois_full_text"] =~ /RIPE/
-        out = _whois_query_ripe_ip(lookup_string, out)
-      elsif out["whois_full_text"] =~ /APNIC/
+      if whois_text =~ /RIPE/
+        _log "Querying RIPE"
+        out = _whois_query_ripe_ip(lookup_string) # _whois_query_ripe_ip(lookup_name)
+      
+      elsif whois_text =~ /APNIC/
         _log "using RDAP to query APNIC"
-        rdap_info = _rdap_ip_lookup lookup_string
-        out = out.merge(rdap_info)
-      elsif out["whois_full_text"] =~ /whois.lacnic.net/ || out["whois_full_text"] =~ /cert\@cert\.br/
+        out <<  _rdap_ip_lookup(lookup_string) # add our hash
+        
+      elsif whois_text =~ /whois.lacnic.net/ || whois_text =~ /cert\@cert\.br/
         _log "using RDAP to query LACNIC"
-        rdap_info = _rdap_ip_lookup lookup_string
-        out = out.merge(rdap_info) if rdap_info
-      elsif out["whois_full_text"] =~ /AFRINIC/
+        out <<  _rdap_ip_lookup(lookup_string) # add our hash
+        
+      elsif whois_text =~ /AFRINIC/
         _log "using RDAP to query AFRINIC"
-        rdap_info = _rdap_ip_lookup lookup_string
-        out = out.merge(rdap_info)
+        out << _rdap_ip_lookup(lookup_string) # add our hash
+
       else # Default to ARIN
-        out = _whois_query_arin_ip(lookup_string, out)
+        out << _whois_query_arin_ip(lookup_string) # add our hash
+
       end
+
     else
-      _parse_nameservers(parser,out)
-      _parse_contacts(parser,out)
+      hash = {}
+      hash["contacts"] = _parse_contacts(parser)
+      hash["nameservers"] = _parse_nameservers(parser)
+      out << hash
     end
 
-  out
+    # add in whois text for all 
+    out.each do |hash|
+      hash["whois_full_text"] = whois_text
+    end
+
+  out # NOTE THAT THIS IS AN ARRAY
   end
 
   # Returns an array of netblocks related to this org
@@ -137,6 +156,12 @@ module Whois
   out
   end
 
+  def range_to_cidrs(lower, upper)
+    ip_range = IPRanger::IPRange.new(lower, upper)
+    cidrs = ip_range.cidrs
+  cidrs
+  end
+
   private  # helper methods for parsing
 
   # use RDAP to query an IP
@@ -174,7 +199,8 @@ module Whois
       "end_address" => "#{json["endAddress"]}",
       "cidr" => "#{cidr_length}",
       "description" => "#{description}",
-      "block_type" => "#{type}"
+      "block_type" => "#{type}",
+      "extended_rdap" => response
     }
 
   out
@@ -182,66 +208,109 @@ module Whois
 
 
   # returns a hash that can create a netblock
-  def _whois_query_ripe_ip(lookup_string,out={})
-    ripe_uri = "https://stat.ripe.net/data/address-space-hierarchy/data.json?resource=#{lookup_string}/32"
+  def _whois_query_ripe_ip(lookup_string,out=[])
+
+    if lookup_string =~ /:/
+      ripe_uri = "https://stat.ripe.net/data/address-space-hierarchy/data.json?resource=#{lookup_string}/64"
+    else
+      ripe_uri = "https://stat.ripe.net/data/address-space-hierarchy/data.json?resource=#{lookup_string}/32"
+    end
+
     json = JSON.parse(http_get_body(ripe_uri))
+    data = json["data"]
 
     # parse out ranges
-    data = json["data"]
     if data["last_updated"]
       range = data["last_updated"].first["ip_space"]
       start_address = range.split("/").first.strip
       cidr = range.split("/").last.strip
       exact = false
-    elsif !data["exact"].empty? # RIPE
+    elsif !data["exact"].empty? && data["more_specific"].empty?  # RIPE
       range = data["exact"].first["inetnum"]
       start_address = range.split("-").first.strip
       end_address = range.split("-").last.strip
       netname = data["exact"].first["netname"]
       org = data["exact"].first["org"]
       exact = true 
-    elsif !data["more_specific"].empty? && data["more_specific"].first["inetnum"] # RIPE
+    elsif !data["more_specific"].empty? && data["more_specific"].first["inetnum"] # IPv4 CASE
       range = data["more_specific"].first["inetnum"]
       start_address = range.split("-").first.strip
       end_address = range.split("-").last.strip
       netname = data["more_specific"].first["netname"]
       org = data["more_specific"].first["org"]
       exact = false
-    elsif !data["less_specific"].empty? && data["less_specific"].first["inetnum"] # RIPE
+    elsif !data["more_specific"].empty? && data["more_specific"].first["inet6num"] # IPV6 CASE!
+      range = data["more_specific"].first["inet6num"]
+      start_address = range.split("/").first.strip
+      cidr = range.split("/").last.strip
+      netname = data["more_specific"].first["netname"]
+      org = data["more_specific"].first["org"]
+      exact = false
+    elsif !data["less_specific"].empty? && data["less_specific"].first["inetnum"] # IPv4 CASE
       range = data["less_specific"].first["inetnum"]
       start_address = range.split("-").first.strip
       end_address = range.split("-").last.strip
       netname = data["less_specific"].first["netname"]
       org = data["less_specific"].first["org"]
       exact = false
+    elsif !data["less_specific"].empty? && data["less_specific"].first["inet6num"] # IPV6 CASE!
+      range = data["less_specific"].first["inet6num"]
+      start_address = range.split("/").first.strip
+      cidr = range.split("/").last.strip
+      netname = data["less_specific"].first["netname"]
+      org = data["less_specific"].first["org"]
+      exact = false
     else
-      _log_error "Unknown response , unable to continue"
+      _log_error "Unknown response, unable to continue"
       _log "Got: #{data}"
-      return nil
+      return []
     end
 
     # parse out description
     begin
-      less_specific_hash = data["less_specific"]
-      if less_specific_hash && less_specific_hash.first["descr"]
-        _log "less_specific_hash: #{less_specific_hash}"
-        description = less_specific_hash.first["descr"]
+
+      #more_specific_hash = data["more_specific"]
+      #less_specific_hash = data["less_specific"]
+      #if more_specific_hash
+      #  if more_specific_hash && more_specific_hash["descr"]
+      #    _log "more specific hash: #{more_specific_hash}"
+      #    description = more_specific_hash.first["descr"]
+      ##  end
+      #elsif less_specific_hash
+      #  if less_specific_hash && less_specific_hash["descr"]
+      #    _log "less_specific_hash: #{less_specific_hash}"
+      #    description = less_specific_hash.first["descr"]
+      #  end
+      #end
+
+      # convert the range to cidr format
+      unless cidr 
+        cidrs = range_to_cidrs(start_address, end_address).map{|x| x.to_cidr }
+        raise "Multiple CIDRs available!!!" if cidrs.count > 1
       end
 
-      out = out.merge({
-        "exact" => exact,
-        "name" => "#{range}",
-        "start_address" => "#{start_address}",
-        "end_address" => "#{end_address}",
-        "cidr" => "#{cidr}",
-        "rir" => "RIPE",
-        "organization_reference" => "#{netname}".sanitize_unicode,
-        "organization_name" => "#{org}".sanitize_unicode,
-        "provider" => "#{org}".sanitize_unicode
-        })
+      out = []
+      (cidrs || [cidr]).each do  |x|
+        # merge in our details
+        out << {
+          "exact" => exact,
+          "name" => "#{start_address}/#{x.split("/").last}",
+          "start_address" => "#{start_address}",
+          "end_address" => "#{end_address}",
+          # note that this will already be the length if pulled from above, 
+          # split is harmless 
+          "cidr" => "#{x.split("/").last}", 
+          "rir" => "RIPE",
+          #"description" => "#{description}".sanitize_unicode,
+          "organization_reference" => "#{netname}".sanitize_unicode,
+          "organization_name" => "#{org}".sanitize_unicode,
+          "provider" => "#{org}".sanitize_unicode,
+          "multiple_cidrs" => (cidrs || []).count > 1
+        }
+      end
 
     rescue TypeError => e
-      _log_error "PARSING ERROR! Unable to get details from #{less_specific_hash} #{e}"
+      _log_error "PARSING ERROR! Unable to get RIPE details for #{lookup_string} #{e}"
     end
 
   out
@@ -317,9 +386,9 @@ module Whois
     # handle domain contacts
     begin
       _log "Parsing: #{parser.contacts.count} contacts."
-      hash["contacts"] = []
+      contacts = []
       parser.contacts.each do |contact|\
-        hash["contacts"] << {
+        contacts << {
           "name" => "#{contact.name}",
           "email" => "#{contact.email}"
         }
@@ -329,17 +398,17 @@ module Whois
     rescue ::Whois::ParserError => e
       _log_error "Unable to parse attribute: #{e}"
     end
-  hash
+  contacts
   end
 
   def _parse_nameservers(parser, hash={})
     # handle nameservers
+    nameservers = []
     begin
       _log "Parsing: #{parser.nameservers.count} nameservers."
-      hash["nameservers"] = []
       parser.nameservers.each do |nameserver|
         _log "Parsed nameserver: #{nameserver}"
-        hash["nameservers"] << "#{nameserver}"
+        nameservers << "#{nameserver}"
       end
     rescue ::Whois::AttributeNotImplemented => e
       _log_error "Unable to parse attribute: #{e}"
@@ -348,7 +417,7 @@ module Whois
     rescue ::Whois::ResponseIsThrottled => e
       _log_error "Unable to parse attribute: #{e}"
     end
-  hash
+  nameservers
   end
 
 
